% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{baidu_segment}
\alias{baidu_segment}
\title{Use Baidu Cloud word segmentation service}
\usage{
baidu_segment(text, mode = 0)
}
\arguments{
\item{text}{text}

\item{mode}{Using 0, the web content model, or 1 the Baidu query model.}
}
\value{
a structure data of class \code{baidu_segment}
}
\description{
Use Baidu Cloud word segmentation service
}
\examples{
# baidu_segment("张三是个好同志吗？")
}
