% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc-smart-table.R
\name{biocyc_create_smart_table}
\alias{biocyc_create_smart_table}
\title{创建 BioCyc SmartTable}
\usage{
biocyc_create_smart_table(
  session,
  orgid = "ECOLI",
  name = "test",
  type = c("Genes", "Compounds", "Reactions", "Pathways", "Proteins",
    "Transcription-Units", "Attenuators"),
  description = NULL,
  values = c("trpA", "trpB")
)
}
\arguments{
\item{session}{httr::handle对象，表示已建立的BioCyc会话}

\item{orgid}{字符串，生物体数据库标识符，即 orgid（如 "ECOLI"）}

\item{name}{字符串，SmartTable的名称（可选）}

\item{type}{字符串，第一列对象所属的类，如 'Genes'、'Compounds'。默认 "Genes"}

\item{description}{字符串，SmartTable的描述（可选）}

\item{values}{字符串向量，取值}
}
\value{
字符串，创建的SmartTable的 ID
}
\description{
此函数用于创建一个新的 BioCyc SmartTable
}
\details{
For json input files, parameters are encoded as part of the file format.

\strong{JSON Structure for SmartTable Data Retrieval and Creation}

JSON file structure:
\itemize{
\item name (optional) - name field of SmartTable
\item description (optional) - description field of SmartTable
}

Contents of the short form, which is for creating a single column of values only:
\itemize{
\item type: an object class that values will be coerced into for the given database into the SmartTable
\item pgdb: an org id for the database in which SmartTable objects will reside and be coerced with the given type
\item values: an array of strings, which will be coerced into objects for the SmartTable in a single column
}

Contents of the long form, which is for creating multiple columns of values:
\itemize{
\item columns: a list (JSON array) of columns, each of which is:
\item name (optional): a string that will be the name of the column
\item type: an object class that values will be coerced into for the SmartTable
\item rows: a JSON array, each element is a dictionary mapping column ids to values. Each value corresponds to one cell in the SmartTable and contains:
\item value: a string, number, or {frameid: , pgdb: } which will be coerced into an object for the SmartTable
}

JSON examples:

\if{html}{\out{<div class="sourceCode json">}}\preformatted{// Short form
\{"name": "sample short-form group",
"description": "sample short-form description",
"pgdb": "ECOLI",
"type": "Genes",
"values": ["trpA", "trpB"]
\}

// long form (2 columns)
\{"name": "sample long-form group",
 "description": "This is a longer form of JSON used with a SmartTable.",
 "columns": [\{"name": "Gene", "type": "Genes"\}, \{"name": "Expression Level"\}]
 "rows": [\{\{"frameid": "EG11204", "pgdb": "ECOLI"\}, "1.3"\}, 
          \{\{"frameid": "EG11205", "pgdb": "ECOLI"\}, "-2.2"\}]
\}
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
session <- establish_biocyc_session("your_email_address", "your_password")
smart_table_id <- create_smart_table(session, '/path/to/data.json')
print(paste("创建的SmartTable ID:", smart_table_id))
}
}
