% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc-get-all.R
\name{biocyc_get_all}
\alias{biocyc_get_all}
\title{获取所有数据}
\usage{
biocyc_get_all(session, type = "pathways", orgid = "ECOLI")
}
\arguments{
\item{session}{httr::handle对象，表示已建立的BioCyc会话}

\item{type}{字符串，数据类型（如 "pathways", "genes"，详见 details}

\item{orgid}{字符串，生物体数据库标识符（如 "ECOLI"），获取 organism 数据时可以不填}
}
\value{
数据框，包含所有数据
}
\description{
此函数用于获取指定生物体数据库中的所有数据
}
\details{
\code{type} 参数一共支持 23 种类型，分别是（不区分大小写）：
\enumerate{
\item "ATTENUATORS": 衰减子
\item "COMPOUNDS": 化合物
\item "GENES": 基因
\item "ORGANISMS": 物种
\item "PATHWAYS": 通路
\item "PROMOTERS": 启动子
\item "PROTEINS": 全部蛋白质，包括多肽和蛋白质复合体
\item "POLYPEPTIDES": 多肽
\item "PROTEIN-COMPLEXES": 蛋白质复合物
\item "ENZYMES": 酶
\item "RIBOSOMAL-PROTEINS": 核糖体蛋白质
\item "TRANSCRIPTION-FACTORS": 转录因子
\item "TRANSPORTERS": 转运蛋白
\item "CYTOSOLIC-PROTEINS": 细胞质蛋白质
\item "MEMBRANE-PROTEINS": 膜蛋白质
\item "PERIPLASMIC-PROTEINS": 周质蛋白质
\item "PUBLICATIONS": 出版物
\item "REACTIONS": 反应
\item "RIBOSWITCHES": 核糖开关
\item "RNAS": RNA
\item "TERMINATORS": 终止子
\item "TRANSCRIPTION-FACTOR-BINDING-SITES": 转录因子结合位点
\item "TRANSCRIPTION-UNITS": 转录单位
}
}
