% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc-session.R
\name{biocyc_session}
\alias{biocyc_session}
\title{建立BioCyc会话}
\usage{
biocyc_session(
  email = readline("请输入您的BioCyc注册邮箱: "),
  password = readline("请输入您的BioCyc密码: "),
  base_url = "https://websvc.biocyc.org/"
)
}
\arguments{
\item{email}{字符串，BioCyc账户的电子邮箱}

\item{password}{字符串，BioCyc账户的密码}

\item{base_url}{字符串，BioCyc API的基础URL，默认为"https://websvc.biocyc.org/"}
}
\value{
httr::handle对象，表示已建立的会话
}
\description{
此函数用于建立与BioCyc网站的会话连接。
}
\examples{
\dontrun{
session <- establish_biocyc_session("your_email@example.com", "your_password")
}
}
