% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_specific_primer.R
\name{design_specific_primer}
\alias{design_specific_primer}
\title{Design specific primer}
\usage{
design_specific_primer(
  fasta_file,
  minLength = 20,
  maxLength = 25,
  minProductSize = 150,
  maxProductSize = 500,
  verbose = FALSE
)
}
\arguments{
\item{fasta_file}{fasta format}

\item{minLength}{primer min length}

\item{maxLength}{primer max length}

\item{minProductSize}{product min length}

\item{maxProductSize}{product max length}

\item{verbose}{if TRUE, will print progress during the run}
}
\value{
primers
}
\description{
Design specific primer
}
