% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-utils.R
\name{DT_table}
\alias{DT_table}
\title{A table generated by \code{DT::datatable()} with button to export data}
\usage{
DT_table(df)
}
\arguments{
\item{df}{data frame}
}
\value{
a datatable html widget
}
\description{
A table generated by \code{DT::datatable()} with button to export data
}
\examples{
DT_table(head(iris))
}
