% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix.R
\name{extract_from_hist_graph}
\alias{extract_from_hist_graph}
\title{Filter data.frame by graph nodes}
\usage{
extract_from_hist_graph(M = NULL, g = NULL)
}
\arguments{
\item{M}{a data.frame}

\item{g}{a igraph object}
}
\description{
Filter data.frame by graph nodes
}
