% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix-plot.R
\name{four_dimension_barplot}
\alias{four_dimension_barplot}
\title{Show four tags of a bibliometrix data.frame}
\usage{
four_dimension_barplot(M, tags = c("AU", "AU_CO_NR", "AU_UN_NR", "J9"))
}
\arguments{
\item{M}{a bibliometrix data.frame}

\item{tags}{default is \code{c("AU","AU_CO_NR","AU_UN_NR","J9")}}
}
\description{
Show four tags of a bibliometrix data.frame
}
