% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citation.R
\name{get_doi_record}
\alias{get_doi_record}
\title{从 doi.org 获取文献引用信息}
\usage{
get_doi_record(doi)
}
\arguments{
\item{doi}{字符串，文献的 DOI}
}
\value{
包含文献引用信息的列表
}
\description{
从 doi.org 获取文献引用信息
}
\examples{
\dontrun{
record = get_doi_record("10.1038/nature12373")
print(record)
}
}
