% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{get_gpt4_response}
\alias{get_gpt4_response}
\title{获取 GPT-4 响应}
\usage{
get_gpt4_response(
  user_prompt = NULL,
  system_prompt = "You are a helpful assistant.",
  model = "gpt-4o-mini",
  api_key = Sys.getenv("CHATANYWHERE_API_KEY"),
  base_url = "https://api.chatanywhere.tech/v1",
  output_format = "json_object"
)
}
\arguments{
\item{user_prompt}{字符串，用户输入的提示。}

\item{system_prompt}{字符串，系统提示，默认为 "You are a helpful assistant."。}

\item{model}{字符串，指定使用的模型，默认为 "gpt-4o-mini"。}

\item{api_key}{字符串，ChatAnywhere API 密钥，默认从环境变量 "CHATANYWHERE_API_KEY" 获取。}

\item{base_url}{字符串，API 的基础 URL，默认为 "https://api.chatanywhere.tech"。}

\item{output_format}{字符串，输出格式，默认为 "json_object"。}
}
\value{
字符串，GPT-4 模型生成的响应内容。
}
\description{
该函数通过 ChatAnywhere API 发送请求并获取 GPT-4 模型的响应。
}
\examples{
\dontrun{
response <- get_gpt4_response("你好，请介绍一下自己。")
print(response)
}

}
