% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix-plot.R
\name{ggpie}
\alias{ggpie}
\title{plot pie chart for table or vector}
\usage{
ggpie(x, sort = TRUE, decreasing = FALSE)
}
\arguments{
\item{x}{a character vector}

\item{sort}{default is TRUE}

\item{decreasing}{default is FALSE}
}
\description{
plot pie chart for table or vector
}
