% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citation.R
\name{insert_citation}
\alias{insert_citation}
\title{Insert Citation}
\usage{
insert_citation(id, bib_file = "reference.bib")
}
\arguments{
\item{id}{文献的 DOI 或 PubMed ID}

\item{bib_file}{BibTeX 文件的路径，默认为项目根目录下的 "reference.bib"}
}
\value{
插入的引用的 BibTeX 引用键
}
\description{
根据 DOI 或 PubMed ID 插入引用
}
