% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-manage.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{Install missing package and retry}
\usage{
install_missing_packages(expr = bookdown::render_book(), retry = 30)
}
\arguments{
\item{expr}{an expression or command, i. e. bookdown::render_book()}

\item{retry}{how many times should I retry. default is 30}
}
\value{
a message or a error that is not a 'packageNotFoundError'
}
\description{
I always encounter package not found error when I run \code{render_book()} for
the first time. Normally, I check the error message and then rerun the command,
and this make me annoy when I have to do it again and again. Therefore,
I create this function to automate this operation.
}
