% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncom-combination.R
\name{is_child}
\alias{is_child}
\title{Test whether community a is the subset of community b}
\usage{
is_child(parent, child, sep = "/", strict = TRUE)
}
\arguments{
\item{parent}{parent community}

\item{child}{child community}

\item{sep}{separator of community string}

\item{strict}{if TRUE only child will be considered}
}
\value{
TRUE or FALSE
}
\description{
Test whether community a is the subset of community b
}
\examples{
# example code
 is_child("A/B/C/D", "A/B/C")
}
