% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos-read.R
\name{read_fast5k}
\alias{read_fast5k}
\title{读取 WebofScience 导出的 Fast5000 文件}
\usage{
read_fast5k(file, remove = "auto")
}
\arguments{
\item{file}{字符串, Fast5000 文件的路径。}

\item{remove}{字符串,指定是否移除某些数据。默认为 "auto"。}
}
\value{
返回处理后的数据(具体返回类型取决于实现)
}
\description{
读取 WebofScience 导出的 Fast5000 文件
}
\examples{
\dontrun{
data = read_fast5k("path/to/file.fast5")
}
}
