% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnw2rmd.R
\name{rnw2rmd}
\alias{rnw2rmd}
\title{Convert vignettes from Rnw to Rmd}
\usage{
rnw2rmd(
  file,
  output_file = gsub("[Rr]nw$", "Rmd", file),
  output_format = "rmarkdown::html_vignette",
  extra_replacement = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{file}{Path to an *.Rnw file to convert.}

\item{output_file}{Path to write *.Rmd file.}

\item{output_format}{Character string defining the R markdown output format.}

\item{extra_replacement}{a list of file to specify additional replacement.}

\item{overwrite}{Logical, whether existing files should be overwritten.}
}
\value{
Open the output to enable further edition.
}
\description{
This function is derived from \strong{spMisc}.
See also: \url{https://rdrr.io/github/GegznaV/spMisc/src/R/rnw2rmd.R}.
It can convert old Sweave vignettes into R Markdown.
Please do not expect it to do wonders, but to give you a good starting point
for conversions.
}
