% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_add_multiple_push_remote.R
\name{setup_git_remote_url}
\alias{setup_git_remote_url}
\alias{git_add_multiple_push_remote}
\alias{git_delete_push_remote}
\alias{git_show_remote_origin}
\title{Setup git remote push urls}
\usage{
git_add_multiple_push_remote(
  path = ".",
  gitee_user = NULL,
  github_user = NULL,
  repository = basename(R.utils::getAbsolutePath(path)),
  branch = "origin"
)

git_delete_push_remote(
  path = ".",
  gitee_user = NULL,
  github_user = NULL,
  repository = basename(R.utils::getAbsolutePath(path)),
  branch = "origin"
)

git_show_remote_origin(path = ".", branch = "origin")
}
\arguments{
\item{path}{default is current working space}

\item{gitee_user}{default is "soilmicro"}

\item{github_user}{default is "gaospecial"}

\item{repository}{default is the directory name}

\item{branch}{default is "origin"}
}
\value{
a warning or message
}
\description{
Add/delete/check git remote push urls
}
