% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncom-combination.R
\name{sub_community}
\alias{sub_community}
\title{Find n depth children of a synthetic community}
\usage{
sub_community(x, sep = "/", depth = 0)
}
\arguments{
\item{x}{a string like "A/B/C/D"}

\item{sep}{separator}

\item{depth}{if 0 all children will be retured}
}
\value{
string
}
\description{
Find n depth children of a synthetic community
}
\examples{
sub_community("A/B/C/D", depth = 1)
}
