% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix.R
\name{tableTag}
\alias{tableTag}
\title{Count items in characters}
\usage{
tableTag(M, Tag = "CR", sep = ";")
}
\arguments{
\item{M}{a bibliometerix data.frame}

\item{Tag}{name of the tag}

\item{sep}{separator used to link multiple tags in \code{M}}
}
\description{
Count items in characters
}
