% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix.R
\name{tag_by_regex}
\alias{tag_by_regex}
\title{Tag record by regular expression search}
\usage{
tag_by_regex(x, pattern, pattern.names = names(pattern), sep = ";")
}
\arguments{
\item{x}{character}

\item{pattern}{regex}

\item{pattern.names}{pattern names}

\item{sep}{default is ";"}
}
\value{
a table
}
\description{
Tag record by regular expression search
}
