% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citation.R
\name{type_of_id}
\alias{type_of_id}
\title{判断 ID 的类型}
\usage{
type_of_id(id)
}
\arguments{
\item{id}{字符串，输入的 ID。}
}
\value{
字符串，"doi" 或 "pubmed"。
}
\description{
该函数自动判断输入的 ID 是 DOI 还是 PubMed ID。
}
\examples{
type_of_id("10.1000/xyz123")  # 返回 "doi"
type_of_id("12345678")        # 返回 "pubmed"
}
