% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_summary.R
\name{yt_is_playlist}
\alias{yt_is_playlist}
\title{检查 URL 是否为播放列表}
\usage{
yt_is_playlist(url)
}
\arguments{
\item{url}{字符串,YouTube URL}
}
\value{
逻辑值,是否为播放列表
}
\description{
检查 URL 是否为播放列表
}
\keyword{internal}
