% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhipuai-batch-mode.R
\name{zhipuai_batch_results_download}
\alias{zhipuai_batch_results_download}
\title{下载 Batch 批处理任务的结果}
\usage{
zhipuai_batch_results_download(
  batch_id,
  output_path,
  file_names = c("output", "error"),
  api_key = Sys.getenv("ZHIPUAI_API_KEY")
)
}
\arguments{
\item{batch_id}{批处理任务的ID}

\item{output_path}{结果保存的路径}

\item{file_names}{需要下载的文件名,默认c("output", "error")，还可选 “input”}

\item{api_key}{智谱AI的API密钥,默认从环境变量ZHIPUAI_API_KEY中获取}
}
\value{
返回结果保存的路径
}
\description{
此函数用于下载批处理任务的结果
}
