% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhipuai-file.R
\name{zhipuai_file_download}
\alias{zhipuai_file_download}
\title{下载智谱AI平台上的文件}
\usage{
zhipuai_file_download(
  file_id,
  output_path,
  api_key = Sys.getenv("ZHIPUAI_API_KEY")
)
}
\arguments{
\item{file_id}{文件的ID}

\item{output_path}{结果保存的路径}

\item{api_key}{智谱AI的API密钥,默认从环境变量ZHIPUAI_API_KEY中获取}
}
\value{
返回结果保存的路径
}
\description{
此函数用于下载智谱AI平台上的文件
}
\examples{
\dontrun{
zhipuai_download_file("file_id", "output_path")
}
}
